fix(server): el modelo fue creado fuera de la carpeta model
